refactor(typescript): relocated every typescript directory into core directory
