+++ src/core/stores/BooleanStoreObject.ts
