+++ src/core/stores/NumberStoreObject.ts
