+++ src/core/tools/linear/Linear.ts
@@ -1,4 +1,5 @@
 import UIConfig from "./UIConfig";
+import Tool from "../Tool";
 
 class Linear extends Tool {
     constructor() {
