+++ src/core/utils/MathUtils.ts
